Add files via upload
The thesis deals with two types of timeseries forecasting models, which are the artificial neural networks and the statistical models of ARMA, ARIMA and SARIMA. As part of the thesis an application is implemented in the python programming language in order to create predictions based on the above prediction models. More specifically, in the second chapter of the thesis is presented a historical review of artificial neural networks. In addition, the idea, structure, function and architecture of artificial neural networks are described with a special emphasis on the perceptron model. The thesis also presents the perceptron model training algorithm and the multilayer perceptron (MLP). In the third chapter of the thesis the definition of timeseries and its keys features are presented, while analyzing the statistical models ARMA, ARIMA and SARIMA and the ways of selecting the most suitable model for each timeseries. Finally, at the fourth chapter outlines the idea of the application according to which the purpose is to make predictions for selected datasets. For these forecasts four different datasets were used, where in each of them one of the forecasting models that the thesis deals with was applied. The artificial neural networks chosen to make the predictions is the multilayer perceptron (MLP).